reorganize review component and add translation data
