+++ client/src/components/product/reviews/SortOptions.ts
@@ -1,0 +1,7 @@
+// Sort options for review sorting
+export enum SortOption {
+  NEWEST = "newest",
+  OLDEST = "oldest",
+  HIGHEST_RATING = "highest",
+  LOWEST_RATING = "lowest",
+}
