+++ client/src/components/product/reviews/index.ts
@@ -1,0 +1,6 @@
+export { default as StarRating } from "./StarRating";
+export { default as ReviewItem } from "./ReviewItem";
+export { default as ReviewForm } from "./ReviewForm";
+export { default as ReviewPrompt } from "./ReviewPrompt";
+export { default as RatingDistribution } from "./RatingDistribution";
+export { SortOption } from "./SortOptions";
